Moved IO functionality to IO library
