+++ src/IO/Palette.h
